login + admin

+++ OtoYedekParca/OtoYedekParca.Entity/Model.cs
@@ -6,12 +6,12 @@ namespace OtoYedekParca.Entity
     public class Model : IEntity
     {
         [Key]
-        public long ModelId { get; set; }
+        public int ModelId { get; set; }
         public string ModelAdi { get; set; }
-        public long TipId { get; set; }
-        public long MarkaId { get; set; }
+        public int TipId { get; set; }
+        public int MarkaId { get; set; }
         public string? ImagePath { get; set; }
-        public long Siralama { get; set; }
+        public int Siralama { get; set; }
         public bool Pasif { get; set; }
     }
 }
